Logout when token is expired
Spinner on delete

Export opens new tab then begins download

+++ hooks/Logout.tsx
@@ -3,25 +3,29 @@ import { useAuth } from "@/context/AuthContext";
 import { BASE_URL } from "@/lib/constants";
 import { useRouter } from "next/navigation";
 
-const useLogout = () => {
+const useLogout = (tokenExpire: boolean = false) => {
     const { userData, setUserData } = useAuth();
     const router = useRouter();
 
     const logout = async () => {
-        console.log("Running");
-        const res = await fetch(`${BASE_URL}/api/auth/logout/`, {
-            method: "DELETE",
-            headers: {
-                Authorization: `Token ${userData?.token}`,
-            },
-        });
-
-        if (!res.ok) {
-            toast({
-                title: "Alert",
-                description: "Something went wrong!",
-                className: "bg-red-200 border-none",
+        if (!tokenExpire) {
+            const res = await fetch(`${BASE_URL}/api/auth/logout/`, {
+                method: "DELETE",
+                headers: {
+                    Authorization: `Token ${userData?.token}`,
+                },
             });
+
+            if (!res.ok) {
+                toast({
+                    title: "Alert",
+                    description: "Something went wrong!",
+                    className: "bg-red-200 border-none",
+                });
+            } else {
+                setUserData({ token: "" });
+                router.push("/login");
+            }
         } else {
             setUserData({ token: "" });
             router.push("/login");
